chore: address PR follow ups
