+++ kernel/src/parallel/parallel_scan_metadata.rs
@@ -5,7 +5,6 @@ use delta_kernel_derive::internal_api;
 use tracing::{info_span, Span};
 
 use crate::log_replay::{ActionsBatch, ParallelLogReplayProcessor};
-use crate::log_segment::LogSegment;
 use crate::metrics::reporter::emit_scan_metadata_completed;
 use crate::metrics::{MetricId, ScanType};
 use crate::parallel::parallel_phase::ParallelPhase;
@@ -59,8 +58,6 @@ impl SequentialScanMetadata {
         sequential: SequentialPhase<ScanLogReplayProcessor>,
         engine: Arc<dyn Engine>,
         action_schema: SchemaRef,
-        stats_schema: Option<SchemaRef>,
-        partition_schema: Option<SchemaRef>,
     ) -> Self {
         let operation_id = MetricId::new();
         Self {
@@ -70,8 +67,6 @@ impl SequentialScanMetadata {
             deferred_parallel_checkpoint_planning: Some(DeferredParallelCheckpointPlanning {
                 engine,
                 action_schema,
-                stats_schema,
-                partition_schema,
             }),
             span: info_span!("sequential_scan_metadata"),
         }
@@ -103,13 +98,8 @@ impl SequentialScanMetadata {
                 mut processor,
                 files,
             } => {
-                let event = processor.get_metrics().to_event(
-                    operation_id,
-                    ScanType::SequentialPhase,
-                    start.elapsed(),
-                );
                 if let Some(deferred_planning) = deferred_parallel_checkpoint_planning.as_ref() {
-                    if let Some(checkpoint_info) = deferred_planning.finalize(&files)? {
+                    if let Some(checkpoint_info) = deferred_planning.finalize(&processor, &files)? {
                         processor.set_checkpoint_info(
                             deferred_planning.engine.as_ref(),
                             checkpoint_info,
@@ -117,6 +107,11 @@ impl SequentialScanMetadata {
                     }
                 }
 
+                let event = processor.get_metrics().to_event(
+                    operation_id,
+                    ScanType::SequentialPhase,
+                    start.elapsed(),
+                );
                 processor
                     .get_metrics()
                     .log("Sequential scan metadata completed");
@@ -134,31 +129,29 @@ impl SequentialScanMetadata {
             }
         }
     }
+
+    #[cfg(test)]
+    fn checkpoint_info_for_tests(&self) -> &crate::log_segment::CheckpointReadInfo {
+        self.sequential.processor().checkpoint_info()
+    }
 }
 
 struct DeferredParallelCheckpointPlanning {
     engine: Arc<dyn Engine>,
     action_schema: SchemaRef,
-    stats_schema: Option<SchemaRef>,
-    partition_schema: Option<SchemaRef>,
 }
 
 impl DeferredParallelCheckpointPlanning {
     fn finalize(
         &self,
+        processor: &ScanLogReplayProcessor,
         files: &[FileMeta],
     ) -> DeltaResult<Option<crate::log_segment::CheckpointReadInfo>> {
-        if files.is_empty() {
-            return Ok(None);
-        }
-
-        Ok(Some(LogSegment::projected_checkpoint_read_info_from_files(
+        processor.projected_checkpoint_read_info_from_files(
             self.engine.as_ref(),
             self.action_schema.clone(),
-            self.stats_schema.as_deref(),
-            self.partition_schema.as_deref(),
             files,
-        )?))
+        )
     }
 }
 
@@ -332,5 +325,175 @@ impl Iterator for ParallelScanMetadata {
     fn next(&mut self) -> Option<Self::Item> {
         let _guard = self.span.enter();
         self.processor.next()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::arrow::array::{Array, Int64Array, StructArray};
+    use crate::arrow::record_batch::RecordBatch;
+    use crate::engine::arrow_data::ArrowEngineData;
+    use crate::expressions::{column_expr, Expression as Expr};
+    use crate::scan::ScanMetadata;
+    use crate::schema::DataType;
+    use crate::utils::test_utils::load_test_table;
+
+    fn schema_has_add_field(schema: &SchemaRef, field_name: &str) -> bool {
+        let Some(add_field) = schema.field("add") else {
+            return false;
+        };
+        let DataType::Struct(add_struct) = add_field.data_type() else {
+            return false;
+        };
+        add_struct.field(field_name).is_some()
+    }
+
+    fn collect_selected_num_records(
+        rows: &mut Vec<i64>,
+        scan_metadata: ScanMetadata,
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        let (data, selection_vector) = scan_metadata.scan_files.into_parts();
+        let batch: RecordBatch = ArrowEngineData::try_from_engine_data(data)?.into();
+        let stats_parsed = batch
+            .column_by_name("stats_parsed")
+            .expect("stats_parsed column should be present")
+            .as_any()
+            .downcast_ref::<StructArray>()
+            .expect("stats_parsed column should be StructArray");
+        let num_records = stats_parsed
+            .column_by_name("numRecords")
+            .expect("stats_parsed.numRecords should be present")
+            .as_any()
+            .downcast_ref::<Int64Array>()
+            .expect("stats_parsed.numRecords should be Int64Array");
+
+        for (row_index, &selected) in selection_vector.iter().enumerate().take(batch.num_rows()) {
+            if selected {
+                assert!(
+                    !stats_parsed.is_null(row_index),
+                    "stats_parsed should be non-null for selected row {row_index}"
+                );
+                rows.push(num_records.value(row_index));
+            }
+        }
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_parallel_scan_metadata_preserves_stats_parsed_for_single_part_struct_stats_checkpoint(
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        let (engine, snapshot, _tempdir) = load_test_table("v1-single-part-struct-stats-only")?;
+
+        let scan = snapshot
+            .scan_builder()
+            .include_all_stats_columns()
+            .build()?;
+
+        let mut expected_num_records = Vec::new();
+        for scan_metadata in scan.scan_metadata(engine.as_ref())? {
+            collect_selected_num_records(&mut expected_num_records, scan_metadata?)?;
+        }
+        expected_num_records.sort_unstable();
+        assert!(!expected_num_records.is_empty());
+
+        let mut sequential = scan.parallel_scan_metadata(engine)?;
+        let mut actual_num_records = Vec::new();
+        for scan_metadata in sequential.by_ref() {
+            collect_selected_num_records(&mut actual_num_records, scan_metadata?)?;
+        }
+
+        assert!(matches!(
+            sequential.finish()?,
+            AfterSequentialScanMetadata::Done
+        ));
+
+        actual_num_records.sort_unstable();
+        assert_eq!(actual_num_records, expected_num_records);
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_parallel_scan_metadata_plans_partition_values_parsed_for_single_part_checkpoint(
+    ) -> DeltaResult<()> {
+        let (engine, snapshot, _tempdir) = load_test_table("app-txn-checkpoint")?;
+
+        let scan = snapshot
+            .scan_builder()
+            .with_predicate(Arc::new(Expr::eq(
+                column_expr!("modified"),
+                Expr::literal("2021-02-02".to_string()),
+            )))
+            .build()?;
+
+        let sequential = scan.parallel_scan_metadata(engine)?;
+        let checkpoint_info = sequential.checkpoint_info_for_tests();
+
+        assert!(checkpoint_info.has_partition_values_parsed);
+        assert!(
+            schema_has_add_field(
+                &checkpoint_info.checkpoint_read_schema,
+                "partitionValues_parsed"
+            ),
+            "checkpoint read schema should include add.partitionValues_parsed"
+        );
+
+        Ok(())
+    }
+
+    #[test]
+    fn test_parallel_scan_metadata_preserves_stats_parsed_for_v2_sidecar_checkpoints(
+    ) -> Result<(), Box<dyn std::error::Error>> {
+        for table_name in [
+            "v2-parquet-sidecars-struct-stats-only",
+            "v2-json-sidecars-struct-stats-only",
+        ] {
+            let (engine, snapshot, _tempdir) = load_test_table(table_name)?;
+            let scan = snapshot
+                .scan_builder()
+                .include_all_stats_columns()
+                .build()?;
+
+            let mut expected_num_records = Vec::new();
+            for scan_metadata in scan.scan_metadata(engine.as_ref())? {
+                collect_selected_num_records(&mut expected_num_records, scan_metadata?)?;
+            }
+            expected_num_records.sort_unstable();
+            assert!(!expected_num_records.is_empty());
+
+            let mut sequential = scan.parallel_scan_metadata(engine.clone())?;
+            let mut actual_num_records = Vec::new();
+
+            for scan_metadata in sequential.by_ref() {
+                collect_selected_num_records(&mut actual_num_records, scan_metadata?)?;
+            }
+
+            let AfterSequentialScanMetadata::Parallel { state, files } = sequential.finish()?
+            else {
+                panic!("expected parallel phase for {table_name}");
+            };
+
+            assert!(!files.is_empty(), "expected sidecar files for {table_name}");
+            assert!(
+                schema_has_add_field(&state.file_read_schema(), "stats_parsed"),
+                "parallel state file read schema should include add.stats_parsed for {table_name}"
+            );
+
+            let state = Arc::new(*state);
+            let mut parallel = ParallelScanMetadata::try_new(engine, state, files)?;
+            for scan_metadata in parallel.by_ref() {
+                collect_selected_num_records(&mut actual_num_records, scan_metadata?)?;
+            }
+
+            actual_num_records.sort_unstable();
+            assert_eq!(
+                actual_num_records, expected_num_records,
+                "parallel sidecar scan metadata should preserve stats_parsed.numRecords for {table_name}"
+            );
+        }
+
+        Ok(())
     }
 }
